+++ akita/src/api/store-utils.ts
@@ -48,9 +48,11 @@ export interface ResetStoresParams {
   exclude: string[];
 }
 
-
+/**
+ *
+ * @param options
+ */
 export function resetStores(options?: Partial<ResetStoresParams>) {
-
   const defaults: ResetStoresParams = {
     exclude: []
   };